Extract windows.rs, macos.rs in installer
