+++ src/bin/installer/macos.rs
@@ -1,0 +1,3 @@
+pub fn install() -> anyhow::Result<()> {
+    todo!("macOS installer is not yet implemented")
+}
